change enums to enumclasses

+++ Structs/BoardArray.h
@@ -25,7 +25,7 @@ struct BoardArray {
     BoardArray(){
         for(int i=0; i<8; i++){
             for(int j=0; j<8; j++){
-                Pieces[i][j] = ChessPiece::NONE;
+                Pieces[i][j] = static_cast<int>(ChessPiece::NONE);
             }
         }
     }
